requirements.txt
Django==5.0.4
Faker==24.4.0
geopy==2.4.1
pandas==2.2.1
django-allauth~=0.61.1
django-autoslug~=1.9.9
djangorestframework~=3.15.1